IP2-PROJECT/IP2POSTv2-DOCKER-COMPLETE/requirements.txt
Flask
mysql-connector-python
requests
flask-bcrypt
apscheduler
python-magic
clamd